Initial Commit
Includes the initial setup for the Android app and a fragment, still not close to ready.  Also includes the work so far on a generic class and two specific classes for complex number representation, and some tests so far.

At this point, testing has revealed that possibly the equality check is botched, will diagnose further.

A little late for a "first" commit but far better late than never!