Bulk update after a few moments of inspiration.
Reworked generic typing of Complex, moved Exp() and Log() to be static methods in Complex instead of derived classes, and discovered Number.doubleValue().  With that perspective, reinvented Complex.equals() and a few other points.

Currently treating Complex<Number> as the root type for all Complex operations (will move to allow Complex<String> and Complex<List> for less restrictive length on the digits of the underlying number type, eventually)