Implemented Complex.root and Complex.roots methods.
Complex.root tentatively works, will follow up with further testing.

Complex.roots is basically untested, will add that later.  Planning to have it generate a list of
n roots and then just check that each of them to the corresponding power return the original value.

Added Complex.rootNewton as a helper method to determine the real-valued root of the modulus.
The algorithm is based on Newton's method, and seeded with an exponent/logarithm computation of
the root.  Laser accurate on 16.  Should check further.  Will later dig into the computational
efficiency of this seed-then-Newton approach.  Will also check that the integer coercion does
not corrupt anything.

Unit testing revealed that the ComplexDoubleCartesian.Argument method was wrong for the
negative reals, added an extra if block for that condition.

Current unit testing complains about the square root of negative one because it is not affirming the
equality of i and 1*e^(pi/2).  Will open new issue on GitHub with further.