Updated Complex.equals() method
Blocked the .equals() method of the lesser classes and used Complex.equals() to check types and that the magnitude of |A-B| is less than a static variable defined TARGET_PRECISION.  Set precision at 1E-13 for now.  This seems to have alleviated all the other issues.  Now just have some lingering generic type warnings to resolve.