Overhaul of the fragment selection menu
Added a listener interface for selecting topics, a topic selector method to MainActivity, and updated the constructors for the RecyclerView adapter and its views to pass the listener reference along.  Now actually responds to clicks by loading the desired fragment, behavior still not quite optimal.  Should track which fragment is currently active so it won't reload it, and not duplicate fragments in the back stack.