Added TODO for Complex.equals()
In researching other things, I saw a few mentions that Java spec for equals is that equality is transitive and symmetric (see Java 7 API documentation), the current equals() is questionable on the symmetric state in that it accepts lesser classes, and WILL violate transitivity because of its approximately-equals implementation.  Plus it wouldn't pair well with the hashCode() specs.  Conceptually it either needs a separate approximately-equals and equals method, and/or needs a discretized "bin" system for values where anything falling in a raster cell qualifies.  The "bin" system makes sense if we go into a numerical computation algorithm that ensures accuracy to the precision of the T data type, but that's a lofty goal for now.