Implemented a TopicFragment and corresponding activity to better react to topic selection in MainActivity
Current implementation is admittedly a hasty whack-it-with-a-hammer-until-it-works approach.  Poorly designed.  There should be a listener interface defined in MainActivity (or better yet a Topics class) for the topics, and then this interface implemented by the View or Fragment that is paired with it.  Also, it would be suitable for the View or Fragment to have an internal but visible Activity for displaying the fragment, which would be used to launch this any time that we aren't intending dual-pane or expandable/retractable menu display.