Rolling out a Topic class with some generic stuff and better designs to (hopefully) provide a much more robust way for listing and selecting the different topics for the app.
This specifically is the abstract contract for any TopicFragment classes, to ensure that they have the appropriate methods to update for a stored state and have a minimum level of construction, and this onCreate should instead pull from the Topic List that is part of the Topic class.  Will see how things go.  I'll next be actively working on this code, one to build a dummy fragment to check for state persistence and to verify that the creation process is appropriate for both large screen (nested fragments) and small screen (separate fragment activities), and that the transactions don't duplicate a fragment somewhere (currently if you click on Intro or Numerical several times then it adds multiple references to the stack).