Rolling out a Topic class with some generic stuff and better designs to (hopefully) provide a much more robust way for listing and selecting the different topics for the app.
Up to this point, this incorporates broad changes to convert the flow from a manually created array of Topics and where they go, to now use a list of Topics dynamically and invoke new instances of things.  I have tested this in the emulator environment for desired functionality on both large-size displays with nested fragment views, and small-size displays with independent fragment activities.  (In doing so noticed that the click handler was layering fragments on top of each other instead of replacing them, now it is a raw replacement without a back stack).  There currently is not any persistent state tracking functionality, that will probably be the next step, and probably using some dummy fragment.