Reinforcing current progress by establishing solid commit of progress so far
Tried for a bit to move on to tracking persistent states, but can't quite get things working right between fragments and activities and lifecycles, I've gone so far as to establish listener pattern stuff but the listener connections kept getting broken.  Wanting to go ahead and reinforce this part so far so that I can hopefully start from "scratch" on my persistence-tracking efforts without a heap of bad code.