Building a ViewModel to persist state for a sample fragment.
ViewModel now successfully retains state for app lifecycle.  Upgraded several component versions in Gradle to ensure the latest functionality.  DataRetentionDemoFragment and DataRetentionDemoViewModel Javadoc'ed to showcase what worked for me.
Currently state is only retained in the ViewModel and does not store to long-term storage, so the state of the model and fragment renews if the app is closed and re-opened.  That could become a future project, but this satisfies what I was hoping for (that the fragment is consistent after rotations and after minimizing, and even closing).